+++ packages/narrate/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @glissade/narrate
 
+## 0.59.0-pre.1
+
+### Patch Changes
+
+- Updated dependencies
+  - @glissade/scene@0.59.0-pre.1
+  - @glissade/core@0.59.0-pre.1
+
 ## 0.59.0-pre.0
 
 ### Patch Changes